+++ Game/EscapeMenu.cpp
@@ -97,7 +97,7 @@ void EscapeMenu::OnEvent(Sprocket::Event& event)
     }
 }
 
-void EscapeMenu::OnUpdate(float dt)
+void EscapeMenu::OnUpdate(double dt)
 {
     d_displayRenderer.OnUpdate();
